src/Implementations/LinkedLists/328_odd_even_linked_list/odd_even_linked_list.go
package odd_even_linked_list

import . "github.com/bumblebee211996/go-dsa/src/Implementations/LinkedLists"

// Problem Statement: Given a singly linked list, group all odd nodes together followed by the even nodes.
// Please note here we are talking about the node number and not the value in the nodes.
//
// You should try to do it in place. The program should run in O(1) space complexity and O(nodes) time complexity.

// Solution: Using two pointers
// One pointer for the odd positions and the other for the even positions. We join the even position elements
// together and join all the odd position elements together. Once both are done, we join the end of the odd
// elements list to the even list.
func OddEvenList(head *ListNode) *ListNode {
	if head == nil {
		return nil
	}
	odd, even := head, head.Next
	eHead := even
	for even != nil && even.Next != nil {
		odd.Next = even.Next
		odd = odd.Next
		even.Next = odd.Next
		even = even.Next
	}
	odd.Next = eHead
	return head
}
